requirements.txt
boto3==1.29.4
boto3-stubs==1.29.4
botocore==1.32.4
botocore-stubs==1.32.4
jmespath==1.0.1
mypy==1.7.0
mypy-extensions==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
tomli==2.0.1
types-awscrt==0.19.13
types-boto3==1.0.2
types-s3transfer==0.7.0
typing_extensions==4.8.0
urllib3==1.26.18
